Extract signal protection into functions
